Add and wrap KL46Z SLCD driver
I don't know the exact provenance or licence of this driver code.
It was distributed to me as part of a resource bundle
with a dev board on a long-ago training course. I
suspect in large part it's the original Freescale reference code,
but encapsulated into a CPP class by an author
at the training provider.

I've added my own wrapper, which keeps control
of the cursor position and implements Meyer's Singleton.